seeder/requirements.txt
redis
requests
psycopg2-binary 